Change the way to apply the authorize

+++ FontEnd/erp/src/app/modules/human-resources/configuration/approve-status/approve-status.service.ts
@@ -17,6 +17,7 @@ import { FilterModel } from 'src/app/core/models/filter-table.model';
 @Injectable()
 export class ApproveStatusService {
 
+  permission = new PermissionViewModel();
   moduleName = 'HR';
   functionCode = 'HR_CONF_APPROVE_ST';
   url = {
@@ -33,7 +34,8 @@ export class ApproveStatusService {
     private context: SessionContext) { }
 
   getPermission(): PermissionViewModel {
-    return this.context.getPermissionByForm(this.moduleName, this.functionCode);
+    this.permission = this.context.getPermissionByForm(this.moduleName, this.functionCode);
+    return this.permission;
   }
 
   getList(paging: PagingModel, searchText: string) {
@@ -50,6 +52,9 @@ export class ApproveStatusService {
   }
 
   save(model: ApproveStatusViewModel, action: FormActionStatus): Observable<ResponseModel> {
+    if (this.permission.allowInsert === false && this.permission.allowUpdate === false) {
+      return;
+    }
     switch (action) {
       case FormActionStatus.Insert:
         return this.api.insert(this.url.insert, model);
@@ -59,7 +64,9 @@ export class ApproveStatusService {
   }
 
   confirmDelete(itemId: number, version: any): Observable<ResponseModel> {
-
+    if (this.permission.allowDelete === false) {
+      return;
+    }
     return this.dialogService.openConfirmDeleteDialog().pipe(
       switchMap((confirmResponse: boolean) => {
         if (confirmResponse === true) {
@@ -72,6 +79,9 @@ export class ApproveStatusService {
   }
 
   delete(itemId: number, version: any): Observable<ResponseModel> {
+    if (this.permission.allowDelete === false) {
+      return;
+    }
     return this.api.delete(this.url.delete, {id: itemId, rowVersion: version });
   }
 }
